requirements.txt
flask==3.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.0.0
tensorflow==2.14.0
numpy==1.24.3
matplotlib==3.7.3
opencv-python==4.8.0
scikit-learn==1.3.0
h5py==3.9.0
requests==2.31.0
pytesseract==0.3.10
torch==2.1.0
torchvision==0.16.0
